refact: rename publisher class
