+++ lib/pipefy_message.rb
@@ -2,9 +2,8 @@
 
 require_relative "pipefy_message/version"
 require_relative "pipefy_message/broker/aws/configuration"
-require_relative "pipefy_message/broker/aws/sns/publisher"
-require_relative "pipefy_message/base_publisher"
 require_relative "pipefy_message/logging"
+require_relative "pipefy_message/publisher"
 require_relative "pipefy_message/worker"
 require_relative "pipefy_message/providers/broker"
 require_relative "pipefy_message/providers/errors"
