+++ lib/pipefy_message/providers/aws_client/sqs_broker.rb
@@ -5,7 +5,7 @@ module PipefyMessage
   module Providers
     module AwsClient
       # AWS SQS client.
-      class SqsBroker < Broker
+      class SqsBroker < PipefyMessage::Providers::Broker
         attr_reader :config
 
         def initialize(queue_name, opts = {})
@@ -16,7 +16,6 @@ module PipefyMessage
           @sqs = Aws::SQS::Client.new
           queue_url = @sqs.get_queue_url({ queue_name: queue_name }).queue_url
           @poller = Aws::SQS::QueuePoller.new(queue_url, { client: @sqs })
-
         rescue Aws::SQS::Errors::NonExistentQueue, Seahorse::Client::NetworkingError => e
           raise PipefyMessage::Providers::Errors::ResourceError, e.message
         end
