+++ lib/pipefy_message/publisher.rb
