+++ spec/publisher_spec.rb
